add admin page

+++ frontend/src/components/CustomInput/InputSearch.js
@@ -12,6 +12,16 @@ import {
 import {Search} from "@material-ui/icons";
 
 import UserService from "../../repository/axiosUserRepository";
+import RoleService from "../../repository/axiosRoleRepository";
+
+/*
+* props:
+*   data - selected items, Map
+*   roles - search target, boolean
+*   image - show Avatar, boolean
+*   onSelect - what happens on item click, CB
+*   onRemove - what happens on item remove, CB
+* */
 
 class InputSearch extends Component {
     state = {
@@ -19,33 +29,37 @@ class InputSearch extends Component {
         value: '',
     };
 
-    handleChange = async e => {
-        let res = [],
-            value = e.target.value;
-        value.trim().length !== 0 && await UserService.searchUsers(value).then(({data}) => {
-            res = data;
-        });
+    handleChange = e => {
+        let value = e.target.value;
+        this.setState({value: value});
+        !value.trim().length && this.setState({results: []});
 
-        this.setState({results: res, value: value})
+        this.props.roles
+            ? value.trim().length && RoleService.search(value).then(({data}) => {
+            this.setState({results: data});
+        })
+            : value.trim().length && UserService.searchUsers(value).then(({data}) => {
+            this.setState({results: data});
+        })
     };
 
-    handleSelect = user => {
-        console.log(this.props.data);
-        this.props.onSelect(user);
+    handleSelect = item => {
+        this.props.onSelect && this.props.onSelect(item);
         this.setState({results: [], value: ''});
     };
 
     handleRemove = key => {
-        this.props.onRemove(key);
+        this.props.onRemove && this.props.onRemove(key);
         this.setState({results: [], value: ''});
     };
 
     render() {
         return (
             <>
-                <div className='bg-light mb-2 p-3 is-rounded'>
+                <div className='bg-light mb-1 p-3 is-rounded'>
                     <FormControl fullWidth>
                         <Input
+                            autoComplete='off'
                             value={this.state.value}
                             onChange={this.handleChange}
                             placeholder='Search...'
@@ -58,12 +72,19 @@ class InputSearch extends Component {
                             }
                         />
                     </FormControl>
-                    <PillList
-                        data={this.props.data}
-                        onClick={this.handleRemove}
-                    />
+                    {
+                        this.props.data
+                        && <PillList
+                            image={this.props.image}
+                            roles={this.props.roles}
+                            data={this.props.data}
+                            onClick={this.handleRemove}
+                        />
+                    }
                 </div>
                 <SearchResult
+                    image={this.props.image}
+                    roles={this.props.roles}
                     results={this.state.results}
                     onSelect={this.handleSelect}
                 />
@@ -77,50 +98,64 @@ export default InputSearch;
 
 const SearchResult = props => {
     return (
-        <div className='search-result mb-4 bg-light p-2'>
-            {props.results.map(user => (
-                <Media
-                    className='cursor-pointer'
-                    key={user.id}
-                    user={user}
-                    onClick={() => props.onSelect(user)}
-                >
-                    <Avatar
-                        style={{width: '30px', height: '30px'}}
-                        className='border border-danger mr-2'
-                        src={`data:${user.fileDetails.mimeType};base64,${user.fileDetails.base64}`}
-                    >R</Avatar>
-                    <Media.Body className='control'>
-                        {user.username}
-                    </Media.Body>
-                </Media>
-            ))}
-        </div>
+        <>
+            <div className='position-relative'>
+                <div className='search-result mb-4 bg-light position-absolute w-100'>
+                    {props.results.map(item => (
+                        <Media
+                            className='cursor-pointer p-2'
+                            key={item.id}
+                            item={item}
+                            onClick={() => props.onSelect(item)}
+                        >
+                            {
+                                props.image
+                                && <Avatar
+                                    style={{width: '30px', height: '30px'}}
+                                    className='border border-danger mr-2'
+                                    src={`data:${item.fileDetails.mimeType};base64,${item.fileDetails.base64}`}
+                                >R</Avatar>
+                            }
+                            <Media.Body className='control'>
+                                {props.roles
+                                    ? item.roleName
+                                    : item.username}
+                            </Media.Body>
+                        </Media>
+                    ))}
+                </div>
+            </div>
+        </>
     );
 };
 
 const PillList = props => {
     return (
         <div className='mt-3'>
-            {[...props.data.values()].map(user => (
+            {[...props.data.values()].map(item => (
                 <Badge
-                    key={user.id}
+                    key={item.id}
                     pill
                     variant="info"
                     className='mr-2 border'
                 >
                     <Media
                         className='cursor-pointer'
-                        user={user}
-                        onClick={() => props.onClick(user.id)}
+                        item={item}
+                        onClick={() => props.onClick(item.id)}
                     >
-                        <Avatar
-                            style={{width: '15px', height: '15px'}}
-                            className='border border-danger mr-2'
-                            src={`data:${user.fileDetails.mimeType};base64,${user.fileDetails.base64}`}
-                        >R</Avatar>
+                        {
+                            props.image
+                            && <Avatar
+                                style={{width: '15px', height: '15px'}}
+                                className='border border-danger mr-2'
+                                src={`data:${item.fileDetails.mimeType};base64,${item.fileDetails.base64}`}
+                            >R</Avatar>
+                        }
                         <Media.Body className='control'>
-                            {user.username}
+                            {props.roles
+                                ? item.roleName
+                                : item.username}
                             <span className='fa fa-close ml-2'/>
                         </Media.Body>
                     </Media>
